requirements.txt
google-cloud-storage>=1.24.1
